VCPKG

+++ Code/ShellExtension/Utility/String.h
@@ -9,9 +9,9 @@ namespace std
 {
     inline std::wstring convert( const std::string& str )
     {
-        int count = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), NULL, 0);
+        int count = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), (int)str.length(), NULL, 0);
         std::wstring wstr(count, 0);
-        MultiByteToWideChar(CP_UTF8, 0, str.c_str(), str.length(), &wstr[0], count);
+        MultiByteToWideChar(CP_UTF8, 0, str.c_str(), (int)str.length(), &wstr[0], count);
         return wstr;
     }
 }
